Update update_colony.py
We wrote some pseudocode and waiting for syntax things that are needed to do some actual coding.  We also set up an internal clock, such that we can see how long our colony is going for, and created simple methods such as make_new_cells and kill_cells using calculus.